+++ modelo/conectar.php
@@ -5,7 +5,7 @@ class Conectar {
         try {
             $conexion = new mysqli("localhost", "root", "", "ja_barbershop");
         } catch (Exception $e) {
-            die('Error' . $e->getMessage());
+            die('Error' . $e->getMessage("error al conectar con la bbdd"));
         }
 
         return $conexion;
